+++ LibCarla/source/carla/trafficmanager/InMemoryMap.h
@@ -24,6 +24,7 @@
 
 #include "carla/trafficmanager/RandomGenerator.h"
 #include "carla/trafficmanager/SimpleWaypoint.h"
+#include "carla/trafficmanager/CachedSimpleWaypoint.h"
 
 namespace carla {
 namespace traffic_manager {
@@ -38,7 +39,7 @@ namespace bgi = boost::geometry::index;
   using SimpleWaypointPtr = std::shared_ptr<SimpleWaypoint>;
   using NodeList = std::vector<SimpleWaypointPtr>;
   using GeoGridId = crd::JuncId;
-  using WorldMap = carla::SharedPtr<cc::Map>;
+  using WorldMap = carla::SharedPtr<const cc::Map>;
 
   using Point3D = bg::model::point<float, 3, bg::cs::cartesian>;
   using Box = bg::model::box<Point3D>;
@@ -69,6 +70,11 @@ namespace bgi = boost::geometry::index;
     InMemoryMap(WorldMap world_map);
     ~InMemoryMap();
 
+    static void Cook(WorldMap world_map, const std::string& path);
+
+    //bool Load(const std::string& filename);
+    bool Load(const std::vector<uint8_t>& content);
+
     /// This method constructs the local map with a resolution of
     /// sampling_resolution.
     void SetUp();
@@ -85,7 +91,13 @@ namespace bgi = boost::geometry::index;
 
     std::string GetMapName();
 
+    const cc::Map& GetMap() const;
+
   private:
+    void Save(const std::string& path);
+
+    void SetUpDenseTopology();
+    void SetUpSpatialTree();
 
     /// This method is used to find and place lane change links.
     void FindAndLinkLaneChange(SimpleWaypointPtr reference_waypoint);