+++ LibCarla/source/carla/trafficmanager/TrackTraffic.h
@@ -32,8 +32,6 @@ private:
     std::unordered_map<ActorId, std::unordered_set<GeoGridId>> actor_to_grids;
     /// Actors currently passing through grids.
     std::unordered_map<GeoGridId, ActorIdSet> grid_to_actors;
-    /// Waypoints currently taken.
-    WaypointIdSet taken_waypoints;
     /// Current hero location.
     cg::Location hero_location = cg::Location(0,0,0);
 
@@ -51,9 +49,8 @@ public:
                                         const std::vector<SimpleWaypointPtr> waypoints);
 
     ActorIdSet GetOverlappingVehicles(ActorId actor_id) const;
-    bool IsWaypointFree(const uint64_t waypoint_id) const;
-    void AddTakenWaypoint(const uint64_t waypoint_id);
-    void ClearTakenWaypoints();
+    bool IsGeoGridFree(const GeoGridId geogrid_id) const;
+    void AddTakenGrid(const GeoGridId geogrid_id, const ActorId actor_id);
 
     void SetHeroLocation(const cg::Location location);
     cg::Location GetHeroLocation() const;