+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Actor/CarlaActor.cpp
@@ -824,6 +824,23 @@ ECarlaServerResponse FVehicleActor::SetActorAutopilot(bool bEnabled, bool bKeepS
   return ECarlaServerResponse::Success;
 }
 
+ECarlaServerResponse FVehicleActor::ShowVehicleDebugTelemetry(bool bEnabled)
+{
+  if (IsDormant())
+  {
+  }
+  else
+  {
+    auto Vehicle = Cast<ACarlaWheeledVehicle>(GetActor());
+    if (Vehicle == nullptr)
+    {
+      return ECarlaServerResponse::NotAVehicle;
+    }
+    Vehicle->ShowDebugTelemetry(bEnabled);
+  }
+  return ECarlaServerResponse::Success;
+}
+
 ECarlaServerResponse FVehicleActor::EnableCarSim(const FString& SimfilePath)
 {
   if (IsDormant())
@@ -1061,7 +1078,7 @@ ECarlaServerResponse FWalkerActor::SetWalkerState(
 
   FTransform NewTransform = Transform;
   NewTransform.SetLocation(NewLocation);
-  SetActorGlobalTransform(NewTransform);
+
   if (IsDormant())
   {
     FWalkerData* WalkerData = GetActorData<FWalkerData>();
@@ -1088,6 +1105,7 @@ ECarlaServerResponse FWalkerActor::SetWalkerState(
     }
     Controller->ApplyWalkerControl(WalkerControl);
   }
+  SetActorGlobalTransform(NewTransform);
   return ECarlaServerResponse::Success;
 }